refactor almost done, please check
