+++ .gitignore
@@ -1,0 +1,5 @@
+dist
+lightning_logs
+saves
+slurm
+.pdf
