+++ code/pyproject.toml
@@ -1,0 +1,19 @@
+[metadata]
+name = "LuttingerWardFunctionalML"
+version = "0.0.1"
+
+[project]
+name = "LuttingerWardFunctionalML"
+version = "0.0.1"
+authors = [
+  { name="Daniel Springer", email="todo_@fill.in" },
+  { name="Julian Stobbe", email="jstobbe@physnet.uni-hamburg.de"}
+]
+description = "Learning the Luttinger-Ward functional of the Hubbard model with neural networks"
+
+[tool.setuptools.packages]
+find = {} 
+
+[build-system]
+requires = ["setuptools"]
+build-backend = "setuptools.build_meta"
