+++ configs/confmod_auto_encoder.json
@@ -2,23 +2,30 @@
     "AUTO_ENCODER_1": {
         "MODEL_NAME": "auto_encoder",
         "MODEL_WRAPER": "model_wraper_ae",
-        "PATH_TRAIN": "/gpfs/data/fs72150/springerd/Projects/LuttingerWard_from_ML/data/U2c0_b10b50_gmax_10x10.hdf5",
+        "PATH_TRAIN": "D:/data_batch2_nPrune.hdf5",
         "DATA_LOADER": "Dataset_ae_split",
         "dtype": "float64",
 
+        "optimizer": "Adam",
         "batch_size": 10,
         "learning_rate": 1e-4,
         "weight_decay": 1e-5,
         "epochs": 150,
 
+        "SGD_weight_decay": 0.0,
+        "SGD_momentum": 0.9,
+        "SGD_dampening": 0.0,
+        "SGD_nesterov": false,
+
         "dropout_in": 0,
         "dropout": 0,
         "activation": "ReLU",
         "with_batchnorm": false,
+        "loss": "MSE",
 
         "in_dim": 200,
         "latent_dim": 14,
-        "n_layers": 1,
+        "n_layers": 3,
 
         "device_type": "gpu",
         "devices": 1,
