+++ configs/confmod_graph_neural_network.json
