+++ train_scripts/train.py
@@ -1,6 +1,6 @@
 #%%
 import torch
-import models
+import code.models.models as models
 import wrapers
 from torch.utils.data import DataLoader
 import load_data
